requirements.txt
anthropic
python-dotenv
pyyaml
tavily-python
python-frontmatter
requests
